sql: partition source planning into modules per type
